+++ src/main/java/com/bookspot/users/infra/token/google/GoogleTokenVerifier.java
@@ -1,4 +1,4 @@
-package com.bookspot.users.infra.token;
+package com.bookspot.users.infra.token.google;
 
 import com.bookspot.global.auth.OAuthClientProperties;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
